separate bids.js into separate testing functions
