+++ validators/bids/obj.js
@@ -1,0 +1,4 @@
+module.exports = {
+  options: {},
+  issues: [],
+}
