+++ validators/bids/reset.js
@@ -1,0 +1,11 @@
+/**
+ * Reset
+ *
+ * Resets the in object data back to original values.
+ */
+
+const reset = BIDS => {
+  BIDS.issues = []
+}
+
+module.exports = reset
